requirements.txt
coverage==4.2
numpy==1.11.0
Pillow==3.2.0
py==1.4.31
pytest==2.9.2
pytest-cov==2.3.1
